Add join back-off
Get in transmission limit for join request using the globalDutyRate.
The limitation is made without summing time on air during the period
to be simple.

Limit for hour 1 is 36s, so select duty rate fo 2^7 which limit to 28s

Limit between hour 1 and 11 is 36s,
divide by 2 the duty rate every 4000s, this limit to 35s

Limit after hour 11 is 8.7s/24h
duty rate of 2^14 which limit to 5.7s/24.